+++ CHANGES.rst
@@ -2,6 +2,8 @@
 --------------------
 - Add delete_obs and delete_var parameters to prepare_for_cellxgene() (#287)
 - GSEA: Revise gsea analysis: Bugfixes, save results into adata, rework plots (#345)
+- Add column exists check to bioutils.pseudobulk_table (#356)
+- Fix main title in receptor-ligand network plot (#358)
 
 Changes to notebooks
 ^^^^^^^^^^^^^^^^^^^^